refactor(pass): extract pass to param
