+++ index.html
@@ -9,6 +9,7 @@
     </title>
     <link rel="stylesheet" href="./bulma@0.9.0/css/bulma.min.css" />
     <link rel="stylesheet" href="./src/style.css" />
+    <script defer src="./src/functions.js" type="module"></script>
     <script defer src="./src/script.js" type="module"></script>
   </head>
